fix CI issues
